Fix BIDMAS op precedence for `+`/`-` ops There is one special case where if there are adjacent `+` and `-` operators in an expression, BIDMAS/BODMAS dictates that the expression should be evaluated linearly.
For example, `2+5-3+6` is evaluated as `((2+5)-3)+6`,  which is equal
to 10. atto currently evaluates this as `(2+5)-(3+6)`, which gives the
incorrect answer of -2. This is because `-` takes a lower precedence
than `+` in atto, where in actual fact, `+` and `-` essentially are on
the same precedence 'tier'.

Another interesting edge case that has been fixed is using a `-` after
another operator to multiply a number by a negative number (eg. `5*-2`),
and other similar expressions. This again was caused generally by the
same issue, but has since been fixed for the majority of cases through
checking whether `-` preceeds another operator token during parsing.

This fix is implemented by removing the `SubtractionExpression` class,
and making `AdditionExpression` additionally parse the `-` token. Child
expression classes are then negated in computed value by using a
`NegationHelperExpression` for expressions after a `-`. This both
retains the functionality of negative number literals (eg. `print -10`),
in addition to typical subtraction (eg. `print 5-2`).

For the multiplication-by-a-negative-number fix, the
`LiteralNegationExpression` class has been introduced at the very top of
the operator precedence order.

There should be minimal impact from this change, since many programs
that contain expressions which are vague in precendence typically make
use of brackets (`(` and `)`) to denote how an expression is evaluated.